+++ DataAccess/Models/Product/UpdateProductModel.cs
@@ -5,11 +5,11 @@ namespace ECommerce.DataAccess.Models.Product
         public int ProductId { get; set; }
         public string? Name { get; set; }
         public string? Description { get; set; }
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
         public int? ManufacturerId { get; set; }
         public int? CategoryId { get; set; }
 
-        public UpdateProductModel(int productId, string? name, string? description, double? price, int? manufacturerId, int? categoryId)
+        public UpdateProductModel(int productId, string? name, string? description, decimal? price, int? manufacturerId, int? categoryId)
         {
             ProductId = productId;
             Name = name;
